% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_peak_position.R
\name{bin_peak_position}
\alias{bin_peak_position}
\title{Assign a peak position to a bin number}
\usage{
bin_peak_position(peak_position, breaks)
}
\arguments{
\item{peak_position}{one peak position (numeric)}

\item{breaks}{vector defining the boundaries of bins}
}
\value{
an integer, the bin number
}
\description{
Assign a peak position to a bin number
}
\examples{
bin_peak_position(5, 1:10)
bin_peak_position(1, 2)

}
