% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_peak_position.R
\name{binarize_bin_number}
\alias{binarize_bin_number}
\title{Binarize bin number}
\usage{
binarize_bin_number(bin_number, total_number_of_bins)
}
\arguments{
\item{bin_number}{a positive integer}

\item{total_number_of_bins}{a positive integer, at least as big as bin_number}
}
\value{
a binary vector with exactly one nonzero entry
}
\description{
Binarize bin number
}
\examples{
binarize_bin_number(3, 10)
binarize_bin_number(10L, 10L)

}
