% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast.R
\name{calc_contrast}
\alias{calc_contrast}
\title{Calculate all contrasts for 8 founder alleles}
\usage{
calc_contrast(grouping, effects)
}
\arguments{
\item{grouping}{a numeric vector, length 8, assigning founder alleles to each of 3 groups: 1, 2, or 3.}

\item{effects}{founder allele effects vector of length 8}
}
\description{
Calculate all contrasts for 8 founder alleles
}
