% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_pleio_sets.R
\name{cluster_pleio_stats}
\alias{cluster_pleio_stats}
\title{Cluster pleiotropy test statistics}
\usage{
cluster_pleio_stats(tib, binarize = FALSE, pleio_threshold = NULL)
}
\arguments{
\item{tib}{a tibble with three columns; the first two are gene ids and the last is pleiotropy test statistics}

\item{binarize}{logical of length one indicating whether to binarize the matrix of pleiotropy test statistics based on pleio_threshold value}

\item{pleio_threshold}{a positive number to distinguish pleiotropy from separate QTL}
}
\description{
Cluster pleiotropy test statistics
}
