% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listcol_to_cols.R
\name{listcol_to_cols}
\alias{listcol_to_cols}
\title{Convert a tibble with one list-column of vectors (of same length) to a tibble without list-column, where each entry of vector (within list-column) is a separate column.}
\usage{
listcol_to_cols(tibble, list_column, another_column)
}
\arguments{
\item{tibble}{a tibble containing one list-column}

\item{list_column}{name of list-column, enclosed with dplyr::quo()}

\item{another_column}{name of another column, enclosed with dplyr::quo()}
}
\value{
a tibble with the list-column replaced with multiple columns. New columns have backticked numbers as names, like `1`
}
\description{
Convert a tibble with one list-column of vectors (of same length) to a tibble without list-column, where each entry of vector (within list-column) is a separate column.
}
\examples{
tibble(a = 1:3, b = c(list(1:2), list(3:4), list(5:6))) \%>\%
listcol_to_cols(list_column = dplyr::quo(b), another_column = dplyr::quo(a))
}
