% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_pleio_sets.R
\name{make_adjacency_matrix}
\alias{make_adjacency_matrix}
\title{Determine adjacency matrix from a n local by m nonlocal traits binary matrix}
\usage{
make_adjacency_matrix(mat)
}
\arguments{
\item{mat}{a binary matrix with rownames and colnames}
}
\description{
Determine adjacency matrix from a n local by m nonlocal traits binary matrix
}
