% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_pleio_sets.R
\name{sort_binary_matrix}
\alias{sort_binary_matrix}
\title{Sort a binary matrix into something close to a block diagonal}
\usage{
sort_binary_matrix(bmat)
}
\arguments{
\item{bmat}{a binary matrix}
}
\description{
Sort a binary matrix into something close to a block diagonal
}
