% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_pleio_sets.R
\name{tibble_to_matrix}
\alias{tibble_to_matrix}
\title{Convert tibble to matrix}
\usage{
tibble_to_matrix(tib, symmetric = FALSE)
}
\arguments{
\item{tib}{a tibble with pleiotropy test results}

\item{symmetric}{logical to indicate if the resulting matrix should be symmetric}
}
\description{
Convert tibble to matrix
}
