% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc2.R
\name{trunc2}
\alias{trunc2}
\title{If the inputted number exceeds `threshold`, return `threshold`; otherwise, return the inputted value.}
\usage{
trunc2(input, threshold)
}
\arguments{
\item{input}{numeric vector to be modified}

\item{threshold}{threshold value}
}
\value{
a numeric vector with length equaling that of `input`
}
\description{
If the inputted number exceeds `threshold`, return `threshold`; otherwise, return the inputted value.
}
\examples{
trunc2(input = 1:10, threshold = 5)
}
